% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conductivity.R
\name{Blanquet}
\alias{Blanquet}
\title{Blanquet routine for ec25 calculation
it fills in some specific rows marked for the Blanquet method}
\usage{
Blanquet(z = dataframeuitMaakKolomMeth)
}
\arguments{
\item{dataframeuitMaakKolomMeth}{the first ec25 calculations start with a dataframe from MaakKolomMeth}
}
\description{
Blanquet routine for ec25 calculation
it fills in some specific rows marked for the Blanquet method
}
