% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateconductivity.R
\name{calculate_conductivity}
\alias{calculate_conductivity}
\title{Calculate the conductivity of groundwater from its ionic composition}
\usage{
calculate_conductivity(
  inputfilename = "data/Table.csv",
  inputstyle = "Stuyfzand",
  outputstyle = "Stuyfzandstyle",
  celcius = 25
)
}
\description{
you will need a data frame with the concentrations of a number of ions
as shown in data/Table.csv
The calculations can use calcium, chloride, Iran, potassium, magnesium, manganese, sodium, ammonium, nitrate, phosphate, zinc, #' bicarbonate, carbonate in milligrams/liter.
Aluminum is used in microgram/liter and the pH is used as such.
The conductivity is calculated at 25°C so the measured conductivity might need some temperature adjustment.
Missing data are assumed to be zero except a missing pH which will be assumed to be 7.
}
