% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conductivity.R
\name{Dunlap}
\alias{Dunlap}
\title{Dunlap routine for ec25 calculation
it fills in some specific rows marked for the Dunlap method}
\usage{
Dunlap(dataframeuitMaakKolomMeth = dataframeuitMaakKolomMeth)
}
\arguments{
\item{dataframeuitMaakKolomMeth}{the second ec25 calculations start with a dataframe from Blanquet}
}
\description{
Dunlap routine for ec25 calculation
it fills in some specific rows marked for the Dunlap method
}
