% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateconductivity.R
\name{LoadFileInVariable}
\alias{LoadFileInVariable}
\title{loads an RData file use variablename<-LoadFileInVariable(filename)}
\usage{
LoadFileInVariable(fileName)
}
\description{
loads an RData file use variablename<-LoadFileInVariable(filename)
}
