% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conductivity.R
\name{Logan}
\alias{Logan}
\title{Logan routine for ec25 calculation
it fills in some specific rows marked for the Logan method}
\usage{
Logan(dataframeuitMaakKolomMeth = dataframeuitMaakKolomMeth)
}
\arguments{
\item{dataframeuitMaakKolomMeth}{the third ec25 calculations start with a dataframe from Dunlap}
}
\description{
Logan routine for ec25 calculation
it fills in some specific rows marked for the Logan method
}
