% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaakKolomMeth.R
\name{MaakKolomMeth}
\alias{MaakKolomMeth}
\title{Prepare for the conductivity calculations by selecting the proper method}
\usage{
MaakKolomMeth(
  LMM_broad_input_dataframe = LMM_broad_input_dataframe,
  celcius = celcius,
  add_bicarbonate = add_bicarbonate,
  add_phosphate = add_phosphate
)
}
\description{
Prepare for the conductivity calculations by selecting the proper method
}
