% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conductivity.R
\name{McNeal}
\alias{McNeal}
\title{McNeal routine for ec25 calculation
it fills in some specific rows marked for the McNeal method}
\usage{
McNeal(dataframeuitMaakKolomMeth = dataframeuitMaakKolomMeth)
}
\arguments{
\item{dataframeuitMaakKolomMeth}{the forth ec25 calculations start with a dataframe from Logan

McNeal routine voor ec25 berekening}
}
\description{
McNeal routine for ec25 calculation
it fills in some specific rows marked for the McNeal method
}
