% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateconductivity.R
\name{ReadListPointComma}
\alias{ReadListPointComma}
\title{Reads a list from a ; separated data file made by WriteListPointComma or Microsoft Excel}
\usage{
ReadListPointComma(filename = "WriteListPointComma.csv")
}
\description{
Reads a list from a ; separated data file made by WriteListPointComma or Microsoft Excel
}
