% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conductivity.R
\name{Rossum}
\alias{Rossum}
\title{Rossum routine for ec25 calculation
it fills in some specific rows marked for the Rossum method}
\usage{
Rossum(dataframeuitMaakKolomMeth = dataframeuitMaakKolomMeth)
}
\arguments{
\item{dataframeuitMaakKolomMeth}{the fifth ec25 calculations start with a dataframe from McNeal

Rossum volgens Ec_voor_Patrick.docx  voor ec25 berekening}
}
\description{
Rossum routine for ec25 calculation
it fills in some specific rows marked for the Rossum method
}
