% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateconductivity.R
\name{WriteListPointComma}
\alias{WriteListPointComma}
\title{Writes a list to a ; separated data file easy to read by WriteListPointComma or Microsoft Excel}
\usage{
WriteListPointComma(inlist, filename = "WriteListPointComma.csv")
}
\description{
Writes a list to a ; separated data file easy to read by WriteListPointComma or Microsoft Excel
}
