% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conductivity.R
\name{calculate_conductivity}
\alias{calculate_conductivity}
\title{Calculate the conductivity of groundwater from its ionic composition}
\usage{
calculate_conductivity(
  inputfilename = "data/input_groundwaterconductivity.rda",
  inputstyle = "Stuyfzand",
  outputstyle = "Stuyfzandstyle",
  celcius = 25
)
}
\arguments{
\item{inputfilename}{The name of the input file in data}

\item{inputstyle}{The layout of the input file like column names etc}

\item{outputstyle}{The layout of the output file}

\item{celcius}{The temperature of the measured conductivity}
}
\value{
A dataframe with the calculated conductivities
}
\description{
you will need a data frame with the concentrations of a number of ions
as shown in data/StuyfzandTable31.csv
The calculations can use calcium, chloride, Iran, potassium, magnesium, manganese, sodium, ammonium, nitrate, phosphate, zinc, #' bicarbonate, carbonate in milligrams/liter.
Aluminum is used in microgram/liter and the pH is used as such.
The conductivity is calculated at 25°C so the measured conductivity might need some temperature adjustment.
Missing data are assumed to be zero except a missing pH which will be assumed to be 7.
myoutputdataframe<-calculate_conductivity(inputfilename="data/input_groundwaterconductivity.rda",inputstyle = "Stuyfzand",outputstyle = "minimal", celcius = 25)
}
