% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadFileInVariable.R
\name{LoadFileInVariable}
\alias{LoadFileInVariable}
\title{Loads an .rda data file into a variable
myTable<-LoadFileInVariable("data/Table.rda")}
\usage{
LoadFileInVariable(filename)
}
\arguments{
\item{filename}{Input datafile like data/Table.rda}
}
\description{
Loads an .rda data file into a variable
myTable<-LoadFileInVariable("data/Table.rda")
}
