% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conductivity.R
\name{MaakKolomMeth}
\alias{MaakKolomMeth}
\title{Prepare for the conductivity calculations by selecting the proper method}
\usage{
MaakKolomMeth(
  LMM_broad_input_groundwaterconductivity = LMM_broad_input_groundwaterconductivity,
  celcius = celcius,
  add_bicarbonate = add_bicarbonate,
  add_phosphate = add_phosphate
)
}
\arguments{
\item{LMM_broad_input_groundwaterconductivity}{your input file was converted to the standard inputfile used here}

\item{celcius}{temperature numerical}

\item{add_bicarbonate}{is TRUE when bicarbonate can be estimated from the ion balance}

\item{add_phosphate}{ist TRUE when total phosphorus is used to estimate phosphate}
}
\description{
Prepare for the conductivity calculations by selecting the proper method
}
