% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadListPointComma.R
\name{ReadListPointComma}
\alias{ReadListPointComma}
\title{Reads a list from a ; separated data file
made by WriteListPointComma or Microsoft Excel}
\usage{
ReadListPointComma(filename = "WriteListPointComma.csv")
}
\arguments{
\item{output}{The output dataframe from the file "data/Table.csv"}
}
\description{
Reads a list from a ; separated data file
made by WriteListPointComma or Microsoft Excel
}
\examples{
Table <- ReadListPointComma(filename = "data/Table.csv")
}
