% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteListPointComma.R
\name{WriteListPointComma}
\alias{WriteListPointComma}
\title{Writes a list or a dataframe to a ; separated data file
easy to read by WriteListPointComma or Microsoft Excel
WriteListPointComma(Table,filename="data/myTable.csv")}
\usage{
WriteListPointComma(Table, filename = "data/WriteListPointComma.csv")
}
\arguments{
\item{Table}{The input dataframe or list Table}

\item{filename}{The name of the data/Table.csv}
}
\description{
Writes a list or a dataframe to a ; separated data file
easy to read by WriteListPointComma or Microsoft Excel
WriteListPointComma(Table,filename="data/myTable.csv")
}
